requirements.txt
langchain
langchain_community
elasticsearch
openai
pydantic
python-dotenv